API/requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.7
django-debug-toolbar==5.0.1
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Markdown==3.7
msgpack==1.1.0
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.9.0
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.12.2
